Fix term_customizer issues after upgrade to v0.27: BOSA21Q1-935 Anderlecht: server problem with import translation set BOSA21Q1-934 Term Customizer "auto-complete"
Author of original `term_customizer` module is on the way to prepare the module upgrade to v0.27, so I prefer switching to the original gem.
All the changes made by me in our fork for the upgrade are also part of `develop` branch in original gem, + there are more fixes, including the two we faced within mentioned tickets.