BOSA21Q1-937 Fix participatory processes counter shows wrong amount of processes.
Fix participatory processes filters doesn't consider current organization.
After switching to Ransack the `organization` filter became broken, it was ignored by ransack.
Original code for the reference: https://github.com/decidim/decidim/blob/v0.27.0/decidim-participatory_processes/app/cells/decidim/participatory_processes/process_filters_cell.rb#L40-L49

While the issue is in decidim it's not proposed as a fix for them because this component is likely to change in the redesign, so the fix won't be accepted as it is for old design.